reservas-backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysqlclient==2.1.1
psycopg2==2.9.5
PyMySQL==1.0.2
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
